refactor: move one use components to views
