+++ src/views/Problems/Index.vue
@@ -1,0 +1,11 @@
+<template>
+  <v-app id="problem-sets-view">
+    <router-view />
+  </v-app>
+</template>
+
+<script>
+export default {
+  name: "ProblemSetViews"
+};
+</script>
