+++ src/views/SignUp/Index.vue
